Move LokiStack Retrieval into a separate hook
