+++ src/components/TracingPage.tsx
@@ -3,61 +3,21 @@ import { Helmet, HelmetProvider } from 'react-helmet-async';
 import {
   Page,
   PageSection,
-  SelectOptionProps,
   Text,
   TextContent,
   Title,
 } from '@patternfly/react-core';
-import { k8sList } from '@openshift-console/dynamic-plugin-sdk';
 import { LokiDropdown } from './LokiDropdown';
 import { useURLState } from '../hooks/useURLState';
+import { useLokiStack } from '../hooks/useLokiStack';
 
 import './example.css';
 
-export const LokiModel = {
-  kind: 'LokiStack',
-  label: 'LokiStack',
-  labelKey: 'public~LokiStack',
-  labelPlural: 'LokiStacks',
-  labelPluralKey: 'public~LokiStacks',
-  apiGroup: 'loki.grafana.com',
-  apiVersion: 'v1',
-  abbr: 'LS',
-  namespaced: true,
-  plural: 'lokistacks',
-};
-
 export default function TracingPage() {
   const { lokiStack, namespace, setLokiStackInURL } = useURLState();
-  const [lokiList, setLokiList] = React.useState<Array<SelectOptionProps>>([]);
-  React.useEffect(() => {
-    /**
-     * TODO: move the data fetching out of a specific component
-     * Consider using one of the watch resources rather than a static fetch
-     */
-    k8sList({ model: LokiModel, queryParams: [] }).then((list) => {
-      const dropdownOptions: Array<SelectOptionProps> = [];
-      if (Array.isArray(list)) {
-        list.forEach((lokiStack) => {
-          dropdownOptions.push({
-            value:
-              lokiStack.metadata.namespace + ' / ' + lokiStack.metadata.name,
-            children:
-              lokiStack.metadata.namespace + ' / ' + lokiStack.metadata.name,
-          });
-        });
-      } else {
-        list.items.forEach((lokiStack) => {
-          dropdownOptions.push({
-            value: lokiStack.metadata.name,
-            children: lokiStack.metadata.name,
-          });
-        });
-      }
-      setLokiList(dropdownOptions);
-    });
-  }, []);
-  if (!lokiList) {
+  const { lokiStackList } = useLokiStack();
+
+  if (!lokiStackList) {
     return <div>Loading...</div>;
   }
   return (
@@ -75,12 +35,12 @@ export default function TracingPage() {
           <label htmlFor="lokistack-dropdown">Select a LokiStack</label>
           <LokiDropdown
             id="lokistack-dropdown"
-            selectionOptions={lokiList}
+            selectionOptions={lokiStackList}
             selectedLokiList={lokiStack}
             selectedNamespace={namespace}
             setLokiList={setLokiStackInURL}
           />
-          {lokiList.find(
+          {lokiStackList.find(
             (listItem) => listItem.value === namespace + ' / ' + lokiStack,
           ) && (
             <TextContent>
